create index.js for models/ to consolidate imports
