+++ src/js/views/05_overtime.js
@@ -4,9 +4,7 @@ import { View, ViewTable } from './view_class.js'
 import Table from '../components/table/table.js';
 import Form from '../components/form/form.js';
 
-import FundLookupTable from '../models/fund_lookup_table.js';
-
-import Services from '../models/services.js';
+import { FundLookupTable, Services } from '../models/';
 import { unformatCurrency } from '../utils/common_utils.js';
 
 export class OvertimeView extends View {
