+++ src/js/views/08_summary.js
@@ -25,6 +25,8 @@ export function compareToTarget(){
 
 function returnToWelcome() {
     const welcome = new WelcomeView();
+    const left = document.getElementById('option1');
+    console.log(left.l)
     welcome.visit();
 };
