+++ src/js/views/view_logic.js
@@ -8,9 +8,7 @@ import NonPersonnelView from './06_nonpersonnel.js';
 import InitiativesView from './07_new_initiatives.js';
 import SummaryView from './08_summary.js';
 
-import FundLookupTable from '../models/fund_lookup_table.js';
-import CurrentFund from '../models/current_fund.js';
-import CurrentPage from '../models/current_page.js';
+import { FundLookupTable, CurrentFund, CurrentPage } from '../models/';
 import { FISCAL_YEAR } from '../init.js';
 
 export function initializePages() {
